using RetestItems for tests
