+++ test/constraints/form/form_sample_list_tests.jl
@@ -1,4 +1,5 @@
 @testitem "PointMassFormConstraint" begin
+    using DomainSets, StableRNGs, DomainSets, Distributions, Random, LinearAlgebra
     import RxInfer: SampleListFormConstraint, is_point_mass_form_constraint, constrain_form
     @testset "is_point_mass_form_constraint" begin
         @test !is_point_mass_form_constraint(SampleListFormConstraint(100))
