+++ test/models/statespace/hgf_tests.jl
@@ -1,75 +1,71 @@
-module RxInferModelsHGFTest
+@testitem "Hierarchical Gaussian Filter" begin
+    using RxInfer, BenchmarkTools, Random, Plots, Dates, LinearAlgebra, StableRNGs
 
-using Test, InteractiveUtils
-using RxInfer, BenchmarkTools, Random, Plots, Dates, LinearAlgebra, StableRNGs
+    # `include(test/utiltests.jl)`
+    include(joinpath(@__DIR__, "..", "..", "utiltests.jl"))
 
-# `include(test/utiltests.jl)`
-include(joinpath(@__DIR__, "..", "..", "utiltests.jl"))
+    # We create a single-time step of corresponding state-space process to
+    # perform online learning (filtering)
+    @model function hgf(real_k, real_w, z_variance, y_variance)
 
-# We create a single-time step of corresponding state-space process to
-# perform online learning (filtering)
-@model function hgf(real_k, real_w, z_variance, y_variance)
+        # Priors from previous time step for `z`
+        zt_min_mean = datavar(Float64)
+        zt_min_var  = datavar(Float64)
 
-    # Priors from previous time step for `z`
-    zt_min_mean = datavar(Float64)
-    zt_min_var  = datavar(Float64)
+        # Priors from previous time step for `x`
+        xt_min_mean = datavar(Float64)
+        xt_min_var  = datavar(Float64)
 
-    # Priors from previous time step for `x`
-    xt_min_mean = datavar(Float64)
-    xt_min_var  = datavar(Float64)
+        zt_min ~ Normal(mean = zt_min_mean, var = zt_min_var)
+        xt_min ~ Normal(mean = xt_min_mean, var = xt_min_var)
 
-    zt_min ~ Normal(mean = zt_min_mean, var = zt_min_var)
-    xt_min ~ Normal(mean = xt_min_mean, var = xt_min_var)
+        # Higher layer is modelled as a random walk 
+        zt ~ Normal(mean = zt_min, var = z_variance)
 
-    # Higher layer is modelled as a random walk 
-    zt ~ Normal(mean = zt_min, var = z_variance)
+        # Lower layer is modelled with `GCV` node
+        gcvnode, xt ~ GCV(xt_min, zt, real_k, real_w)
 
-    # Lower layer is modelled with `GCV` node
-    gcvnode, xt ~ GCV(xt_min, zt, real_k, real_w)
+        # Noisy observations 
+        y = datavar(Float64)
+        y ~ Normal(mean = xt, var = y_variance)
 
-    # Noisy observations 
-    y = datavar(Float64)
-    y ~ Normal(mean = xt, var = y_variance)
-
-    return gcvnode
-end
-
-@constraints function hgfconstraints()
-    q(xt, zt, xt_min) = q(xt, xt_min)q(zt)
-end
-
-@meta function hgfmeta()
-    # Lets use 31 approximation points in the Gauss Hermite cubature approximation method
-    GCV(xt_min, xt, zt) -> GCVMetadata(GaussHermiteCubature(31))
-end
-
-## Inference definition
-function hgf_online_inference(data, vmp_iters, real_k, real_w, z_variance, y_variance)
-    autoupdates = @autoupdates begin
-        zt_min_mean, zt_min_var = mean_var(q(zt))
-        xt_min_mean, xt_min_var = mean_var(q(xt))
+        return gcvnode
     end
 
-    return rxinference(
-        model         = hgf(real_k, real_w, z_variance, y_variance),
-        constraints   = hgfconstraints(),
-        meta          = hgfmeta(),
-        data          = (y = data,),
-        autoupdates   = autoupdates,
-        keephistory   = length(data),
-        historyvars   = (xt = KeepLast(), zt = KeepLast()),
-        initmarginals = (zt = NormalMeanVariance(0.0, 5.0), xt = NormalMeanVariance(0.0, 5.0)),
-        iterations    = vmp_iters,
-        free_energy   = true,
-        autostart     = true,
-        callbacks     = (after_model_creation = (model, returnval) -> begin
-            gcvnode = returnval
-            setmarginal!(gcvnode, :y_x, MvNormalMeanCovariance([0.0, 0.0], [5.0, 5.0]))
-        end,)
-    )
-end
+    @constraints function hgfconstraints()
+        q(xt, zt, xt_min) = q(xt, xt_min)q(zt)
+    end
 
-@testset "Hierarchical Gaussian Filter" begin
+    @meta function hgfmeta()
+        # Lets use 31 approximation points in the Gauss Hermite cubature approximation method
+        GCV(xt_min, xt, zt) -> GCVMetadata(GaussHermiteCubature(31))
+    end
+
+    ## Inference definition
+    function hgf_online_inference(data, vmp_iters, real_k, real_w, z_variance, y_variance)
+        autoupdates = @autoupdates begin
+            zt_min_mean, zt_min_var = mean_var(q(zt))
+            xt_min_mean, xt_min_var = mean_var(q(xt))
+        end
+
+        return rxinference(
+            model         = hgf(real_k, real_w, z_variance, y_variance),
+            constraints   = hgfconstraints(),
+            meta          = hgfmeta(),
+            data          = (y = data,),
+            autoupdates   = autoupdates,
+            keephistory   = length(data),
+            historyvars   = (xt = KeepLast(), zt = KeepLast()),
+            initmarginals = (zt = NormalMeanVariance(0.0, 5.0), xt = NormalMeanVariance(0.0, 5.0)),
+            iterations    = vmp_iters,
+            free_energy   = true,
+            autostart     = true,
+            callbacks     = (after_model_creation = (model, returnval) -> begin
+                gcvnode = returnval
+                setmarginal!(gcvnode, :y_x, MvNormalMeanCovariance([0.0, 0.0], [5.0, 5.0]))
+            end,)
+        )
+    end
 
     ## Data creation
     function generate_data(rng, k, w, zv, yv)
@@ -145,6 +141,4 @@ end
     end
 
     @test_benchmark "models" "hgf" hgf_online_inference($y, $vmp_iters, $real_k, $real_w, $z_variance, $y_variance)
-end
-
 end
