+++ test/models/statespace/hmm_tests.jl
@@ -1,49 +1,45 @@
-module ReactiveMPModelsHMMTest
+@testitem "Hidden Markov Model" begin
+    using BenchmarkTools, Random, Plots, Dates, LinearAlgebra, StableRNGs
 
-using Test, InteractiveUtils
-using RxInfer, BenchmarkTools, Random, Plots, Dates, LinearAlgebra, StableRNGs
+    # `include(test/utiltests.jl)`
+    include(joinpath(@__DIR__, "..", "..", "utiltests.jl"))
 
-# `include(test/utiltests.jl)`
-include(joinpath(@__DIR__, "..", "..", "utiltests.jl"))
+    ## Model definition
+    @model function hidden_markov_model(n)
+        A ~ MatrixDirichlet(ones(3, 3))
+        B ~ MatrixDirichlet([10.0 1.0 1.0; 1.0 10.0 1.0; 1.0 1.0 10.0])
 
-## Model definition
-@model function hidden_markov_model(n)
-    A ~ MatrixDirichlet(ones(3, 3))
-    B ~ MatrixDirichlet([10.0 1.0 1.0; 1.0 10.0 1.0; 1.0 1.0 10.0])
+        s_0 ~ Categorical(fill(1.0 / 3.0, 3))
 
-    s_0 ~ Categorical(fill(1.0 / 3.0, 3))
+        s = randomvar(n)
+        x = datavar(Vector{Float64}, n)
 
-    s = randomvar(n)
-    x = datavar(Vector{Float64}, n)
+        s_prev = s_0
 
-    s_prev = s_0
-
-    for t in 1:n
-        s[t] ~ Transition(s_prev, A)
-        x[t] ~ Transition(s[t], B)
-        s_prev = s[t]
+        for t in 1:n
+            s[t] ~ Transition(s_prev, A)
+            x[t] ~ Transition(s[t], B)
+            s_prev = s[t]
+        end
     end
-end
 
-@constraints function hidden_markov_constraints()
-    q(s, s_0, A, B) = q(s, s_0)q(A)q(B)
-end
+    @constraints function hidden_markov_constraints()
+        q(s, s_0, A, B) = q(s, s_0)q(A)q(B)
+    end
 
-## Inference definition
-function hidden_markov_model_inference(data, vmp_iters)
-    return inference(
-        model = hidden_markov_model(length(data)),
-        constraints = hidden_markov_constraints(),
-        data = (x = data,),
-        options = (limit_stack_depth = 500,),
-        free_energy = true,
-        initmarginals = (A = vague(MatrixDirichlet, 3, 3), B = vague(MatrixDirichlet, 3, 3), s = vague(Categorical, 3)),
-        iterations = vmp_iters,
-        returnvars = (s = KeepEach(), A = KeepEach(), B = KeepEach())
-    )
-end
-
-@testset "Hidden Markov Model" begin
+    ## Inference definition
+    function hidden_markov_model_inference(data, vmp_iters)
+        return inference(
+            model = hidden_markov_model(length(data)),
+            constraints = hidden_markov_constraints(),
+            data = (x = data,),
+            options = (limit_stack_depth = 500,),
+            free_energy = true,
+            initmarginals = (A = vague(MatrixDirichlet, 3, 3), B = vague(MatrixDirichlet, 3, 3), s = vague(Categorical, 3)),
+            iterations = vmp_iters,
+            returnvars = (s = KeepEach(), A = KeepEach(), B = KeepEach())
+        )
+    end
 
     ## Data creation
     function rand_vec(rng, distribution::Categorical)
@@ -106,6 +102,4 @@ end
     end
 
     @test_benchmark "models" "mlgssm" hidden_markov_model_inference($x_data, 20)
-end
-
 end
